+++ src/Provider/Youtube/SignatureDecipher.php
@@ -82,57 +82,6 @@ class SignatureDecipher
     }
 
     /**
-     * decipher a signature with a raw player script
-     *
-     * @deprecated since version 0.7, to be removed in 0.8.
-     *
-     * @param string $decipherScript
-     * @param string $signature
-     * @param Logger $logger
-     *
-     * @return string returns the deciphered signature
-     */
-    public static function decipherSignatureWithRawPlayerScript($decipherScript, $signature, Logger $logger = null)
-    {
-        @trigger_error(__METHOD__ . ' is deprecated since version 0.7, to be removed in 0.8.', E_USER_DEPRECATED);
-
-        // BC: Use NullLogger if no Logger was set
-        if ($logger === null) {
-            $logger = new NullLogger;
-        }
-
-        $opcode = self::extractDecipherOpcode($decipherScript, $logger);
-
-        $decipheredSignature = self::executeSignaturePattern(
-            $opcode['decipherPatterns'],
-            $opcode['deciphers'],
-            $signature,
-            $logger
-        );
-
-        // For debugging
-        $logger->debug(
-            '{method}: Results:',
-            ['method' => __METHOD__]
-        );
-
-        $logger->debug(
-            '{method}: Signature = {signature}',
-            ['method' => __METHOD__, 'signature' => $signature]
-        );
-
-        $logger->debug(
-            '{method}: Deciphered = {decipheredSignature}',
-            ['method' => __METHOD__, 'decipheredSignature' => $decipheredSignature]
-        );
-
-        //file_put_contents("Deciphers".rand(1, 100000).".log", ob_get_contents()); // If you need to debug all video
-
-        //Return signature
-        return $decipheredSignature;
-    }
-
-    /**
      * extract decipher opcode from raw player script
      *
      * @param string $decipherScript
@@ -155,7 +104,7 @@ class SignatureDecipher
 
             return '';
         }
-        
+
         $decipherPatterns = explode('.split("")', $decipherScript);
         unset($decipherPatterns[0]);
         foreach ($decipherPatterns as $value) {
@@ -174,7 +123,7 @@ class SignatureDecipher
                 break;
             }
         }
-        
+
         $logger->debug(
             '{method}: decipherPatterns = {decipherPatterns}',
             ['method' => __METHOD__, 'decipherPatterns' => $decipherPatterns]